Codegen: Refactor functions out of cpu and into c++
